Spicy-13: Simplify and create test
